fetch is broken out into hook, and made a component for front page
