+++ src/hooks/fetch.js
@@ -1,0 +1,32 @@
+import { useEffect, useState } from 'react';
+
+const API_KEY = '830e7f8d39529be36377801e8d5b2393';
+const API_URL = `https://api.themoviedb.org/3/movie/popular?api_key=${API_KEY}&language=en-US&page=1`;
+
+export const useMovies = () => {
+  const [movies, setMovies] = useState([]);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState('');
+
+  useEffect(() => {
+    const fetchMovies = async () => {
+      setLoading(true);
+      try {
+        const response = await fetch(API_URL);
+        if (!response.ok) throw new Error('Failed to fetch');
+        const data = await response.json();
+        setMovies(data.results);
+      } catch (err) {
+        setError(err.message);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchMovies();
+  }, []);
+
+  return { movies, loading, error };
+};
+
+
